Code review feedback: refactor into smaller components, add loading state to storybook.
